Refactor status test (#555)
